Fetching Forms Data const forms = await getForms(); The getForms function is called to retrieve the forms associated with the user. This function likely makes a database query to fetch forms related to the current user.
const totalNumberOfFormCreated = forms?.data?.length || (0 as number);
This line calculates the total number of forms created by the user. If the forms.data is undefined or null, it defaults to 0. The ?. operator is optional chaining, which ensures that the code does not throw an error if forms or forms.data is null or undefined.