Importing Clerk Middleware Utilities The code begins by importing two key utilities from Clerk's server SDK for Next.js:
clerkMiddleware:
This is a function provided by Clerk to create custom middleware for Next.js applications. It gives you access to authentication utilities like auth.protect().

createRouteMatcher:
This is a helper function used to define which routes are public (i.e., accessible without authentication). It allows you to pass an array of route patterns and returns a function that checks if a request matches any of them.