Props (Component Inputs) formId: number: This is the ID of the form being published. It is passed from the parent component and is used to generate a URL for the published form.
open: boolean: This boolean prop controls whether the dialog is open or closed. If open is true, the dialog will be visible; if false, the dialog will be hidden.

onOpenChange: (open: boolean) => void: This is a function passed as a prop from the parent. It is used to change the state of the open prop, toggling the visibility of the dialog. This function should handle the logic for opening and closing the dialog.